requirements.txt
fastapi
uvicorn
psycopg-binary
